GGospelGT/backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.1.2
black==25.1.0
boto3==1.40.23
botocore==1.40.23
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.7
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.114.2
flake8==7.3.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
jmespath==1.0.1
jq==1.10.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.2
mypy==1.17.1
mypy_extensions==1.1.0
numpy==2.3.2
oauthlib==3.3.1
packaging==25.0
pandas==2.3.2
passlib[bcrypt]==1.7.4
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.4.0
pluggy==1.6.0
psutil==7.0.0
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.8.0
pymongo==4.6.0
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-http-client==3.3.7
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pytz==2025.2
requests==2.31.0
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
s3transfer==0.13.1
s5cmd==0.2.0
sendgrid==6.12.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
termii-py==0.1.7
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
watchfiles==1.1.0
Werkzeug==3.1.3
